rendering list of Tiles from API data, moved MoviePage level up
