--- src/features/movies/MoviesPage/MovieList/index.js
@@ -1,14 +1,0 @@
-import Tile from "../../../../common/Tile"
-import { useGetPopularMovies } from "./useGetPopularMovies";
-
-const MovieList = () => {
-    useGetPopularMovies();
-
-    return (
-        <div title="Movies">
-            <Tile />
-        </div>
-    )
-};
-
-export default MovieList;
